DataLoader/requirements.txt
msoffcrypto-tool
openpyxl
pandas
pyodbc
python-dotenv
